Remove `cpp/` and create headers per platform
